requirements.txt
# Core
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3

# Database
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1

# Utils
python-dateutil==2.8.2
pytz==2023.3